# ui/src/main/angular/ui/api/mzTab_m_swagger.yml
swagger: '2.0'
info:
  description: This is the mzTab-M reference implementation and validation API service.
  version: 2.0.0
  title: mzTab-M reference implementation and validation API.
  contact:
    email: nils.hoffmann@isas.de
  license:
    name: Apache 2.0
    url: 'http://www.apache.org/licenses/LICENSE-2.0.html'
host: apps.lifs.isas.de
basePath: /mztabvalidator/rest/v2/
...
schemes:
  - https
...
paths:
...
  /validate:
    post:
      description: >
        Validates an mzTab file in XML or JSON representation and reports
        syntactic, structural, and semantic
        errors.
      operationId: validateMzTabFile
      tags:
        - validate
      consumes:
        - application/json
        - application/xml
      produces:
        - application/json
      parameters:
        - name: mztabfile
          in: body
          description: mzTab file that should be validated.
          required: true
          schema:
            $ref: '#/definitions/MzTab'
        - name: level
          in: query
          description: >-
            The level of errors that should be reported, one of ERROR, WARN,
            INFO.
          required: false
          type: string
          enum:
            - info
            - warn
            - error
          default: info
        - name: maxErrors
          in: query
          description: The maximum number of errors to return.
          required: false
          type: integer
          format: int32
          minimum: 0
          maximum: 500
          default: 100
        - name: semanticValidation
          in: query
          description: >-
            Whether a semantic validation against the default rule set should be
            performed.
          required: false
          type: boolean
          default: false
      responses:
        '200':
          description: Validation Okay
          schema:
            type: array
            default: []
            items:
              $ref: '#/definitions/ValidationMessage'
        '415':
          description: Unsupported content type
        '422':
          description: Invalid input
          schema:
            type: array
            default: []
            items:
              $ref: '#/definitions/ValidationMessage'
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
definitions:
  MzTab:
    type: object
    description: |
        mzTab-M is intended as a reporting standard for quantitative results from metabolomics/lipodomics approaches.
        
        This format is further intended to provide local LIMS systems as well as MS metabolomics repositories a simple way to share and combine basic information.
        
        The mzTab-M format consists of four cross-referenced data tables:
            
            * Metadata (MTD), 
            * Small Molecule (SML), 
            * Small Molecule Feature (SMF) and the 
            * Small Molecule Evidence (SME). 
        
        The MTD and SML tables are mandatory, and for a file to contain any evidence about how molecules were quantified or identified by software, then all four tables must be present. The tables must follow the order MTD, SML, SMF and SME, with a blank line separating each table. 
        
        The structure of each table, in terms of the rows and columns that must be present is tightly specified and formally defined and explained in the mzTab-M specification document.
        mzTab-M files MUST have one Metadata (MTD) section and one Small Molecule (SML) Section. In practice, we expect that most files SHOULD also include one Small Molecule Feature (SMF) section, and one Small Molecule Evidence (SME) Section. 
        Files lacking SMF and SME sections can only present summary data about quantified molecules, without any evidence trail for how those values were derived. It will be left to reading software to determine whether additional validation will be requested such that SMF and SME tables MUST be present.
    required:
      - metadata
      - smallMoleculeSummary
      - smallMoleculeFeature
      - smallMoleculeEvidence
    properties:
      metadata:
        $ref: '#/definitions/Metadata'
      smallMoleculeSummary:
        description: |
            The small molecule section is table-based. The small molecule section MUST always come after the metadata section. All table columns MUST be Tab separated. There MUST NOT be any empty cells; missing values MUST be reported using “null” for columns where Is Nullable = “True”.
            
            Each row of the small molecule section is intended to report one final result to be communicated in terms of a molecule that has been quantified. In many cases, this may be the molecule of biological interest, although in some cases, the final result could be a derivatized form as appropriate – although it is desirable for the database identifier(s) to reference to the biological (non-derivatized) form. In general, different adduct forms would generally be reported in the Small Molecule Feature section.
            
            The order of columns MUST follow the order specified below.

            All columns are MANDATORY except for “opt_” columns.
        type: array
        default: []
        minItems: 1
        items:
          $ref: '#/definitions/SmallMoleculeSummary'
      smallMoleculeFeature:
        description: |
            The small molecule feature section is table-based, representing individual MS regions (generally considered to be the elution profile for all isotopomers formed from a single charge state of a molecule), that have been measured/quantified. However, for approaches that quantify individual isotopomers e.g. stable isotope labelling/flux studies, then each SMF row SHOULD represent a single isotopomer.

            Different adducts or derivatives and different charge states of individual molecules should be reported as separate SMF rows.

            The small molecule feature section MUST always come after the Small Molecule Table. All table columns MUST be Tab separated. There MUST NOT be any empty cells. Missing values MUST be reported using “null”.

            The order of columns MUST follow the order specified below.

            All columns are MANDATORY except for “opt_” columns.
        type: array
        default: []
        items:
          $ref: '#/definitions/SmallMoleculeFeature'
      smallMoleculeEvidence:
        description: |
            The small molecule evidence section is table-based, representing evidence for identifications of small molecules/features, from database search or any other process used to give putative identifications to molecules. In a typical case, each row represents one result from a single search or intepretation of a piece of evidence e.g. a database search with a fragmentation spectrum. Multiple results from a given input data item (e.g. one fragment spectrum) SHOULD share the same value under evidence_input_id.

            The small molecule evidence section MUST always come after the Small Molecule Feature Table. All table columns MUST be Tab separated. There MUST NOT be any empty cells. Missing values MUST be reported using “null”.

            The order of columns MUST follow the order specified below.

            All columns are MANDATORY except for “opt_” columns.
        type: array
        default: []
        items:
          $ref: '#/definitions/SmallMoleculeEvidence'
      comment:
        description: |
            Comment lines can be placed anywhere in an mzTab file. These lines must start with the three-letter code COM and are ignored by most parsers. Empty lines can also occur anywhere in an mzTab file and are ignored.
        type: array
        default: []
        items:
          $ref: '#/definitions/Comment'
  Comment:
    type: object
    description: |
        Comment lines can be placed anywhere in an mzTab file. These lines must start with the three-letter code COM and are ignored by most parsers. Empty lines can also occur anywhere in an mzTab file and are ignored.
    x-mztab-example: |
        COM	This is a comment line
    required:
      - prefix
      - msg
    properties:
      prefix:
        type: string
        enum:
          - COM
        default: COM
      msg:
        type: string
      line_number:
        type: integer
        format: int32
...
  Error:
    type: object
    required:
      - code
      - message
    properties:
      code:
        type: integer
        format: int32
      message:
        type: string
  ValidationMessage:
    type: object
    required:
      - code
      - category
      - message
    properties:
      code:
        type: string
      category:
        enum:
          - format
          - logical
          - cross_check
        default: format
      message_type:
        enum:
          - error
          - warn
          - info
        default: info
      message:
        type: string
      line_number:
        type: integer
        format: int64
...